feature: compute electron window options in shared process instead of main process
