+++ packages/shared-process/src/parts/AppWindow/AppWindow.js
@@ -1,42 +1,13 @@
 import * as DefaultUrl from '../DefaultUrl/DefaultUrl.js'
+import * as GetAppWindowOptions from '../GetAppWindowOptions/GetAppWindowOptions.js'
 import * as ParentIpc from '../ParentIpc/ParentIpc.js'
-import * as Platform from '../Platform/Platform.js'
 import * as Preferences from '../Preferences/Preferences.js'
 import * as Screen from '../Screen/Screen.js'
-
-const getWindowOptions = (preferencs, screenWidth, screenHeight) => {
-  const titleBarPreference = preferencs['window.titleBarStyle']
-  const frame = titleBarPreference !== 'custom'
-  const titleBarStyle = titleBarPreference === 'custom' ? 'hidden' : undefined
-  const zoomLevelPreference = preferencs['window.zoomLevel']
-  const zoomLevel = zoomLevelPreference
-  const windowControlsOverlayPreference = Platform.isWindows && preferencs['window.controlsOverlay.enabled']
-
-  const titleBarOverlay = windowControlsOverlayPreference
-    ? {
-        color: '#1e2324',
-        symbolColor: '#74b1be',
-        height: 29,
-      }
-    : undefined
-  return {
-    y: 0,
-    x: screenWidth - 800,
-    width: 800,
-    height: screenHeight,
-    menu: true,
-    background: '#1e2324',
-    titleBarStyle,
-    frame,
-    zoomLevel,
-    titleBarOverlay,
-  }
-}
 
 export const createAppWindow = async (preferences, parsedArgs, workingDirectory, url = DefaultUrl.defaultUrl) => {
   const screenWidth = await Screen.getWidth()
   const screenHeight = await Screen.getHeight()
-  const windowOptions = getWindowOptions(preferences, screenWidth, screenHeight)
+  const windowOptions = GetAppWindowOptions.getAppWindowOptions(preferences, screenWidth, screenHeight)
   return ParentIpc.invoke('AppWindow.createAppWindow2', windowOptions, parsedArgs, workingDirectory, url)
 }
